refactor: move command types to its own file
This way the `Command` file is much shorter and easier to navigate
